fix: to dict test
